Try to add BATS tests, not so easy
Can't figure out how to pull environment variables into the file that's
initiated by a `run`. Or really any other means. Seems like the contents
of an `@test` are run in a subshell. This seems to not be picking up the
environment beforehand? Iunno.

Something to explore later. Don't want to get frustrated spending too
much time on this if something so basic is not working.

Will check more docs & GH issues later.