Getting hard stuck working on merging scopes
It was a little easier in my head. Running into some very bash-y
problems, some of my own making.

One of which being... type declarations are optional, and when left out,
I did not add an implicit `ANY` type (or equivalent).

There is also just a bit of a difficulty keeping track of the 4 parallel
trees that we're working with...
   - Parent scope
   - Parent AST
   - Child scope
   - Child AST
...particularly with the addition of all the namerefs and whatnot.