Refactor: `scope` -> `symtab`
It's more accurate to describe the table that maps symbols as a symbol
table, rather than a scope. The ambiguous naming was actually making
things difficult for me to sometimes keep track of in my head.

I think this will be both a more accurate representation of what it
actually is, as well as being a bit easier for me to reason about.