Playing around with shellspec, may use in future
I'm still not 100% on `shellspec`. It's a colosal project, and feels
like it will be the right choice to learn eventually.

Currently the native bash syntax of BATS is more helpful. I'm trying to
write many quick tests to ensure core functionality works, and I haven't
broken anything.

That's made more difficult if I need to spend an hour looking up
`shellspec` commands, and learning their way of doing things. The
reporting and native kcov are very nice. It looks like it has the
capacity to do more. Maybe. Though at a significantly increased
timecost.