Add tests for array edge cases & failures
Still pretty poor messaging surrounding when one attempts to declare a
variable within an array. That should not be caught by the pratt parser.
Ideally need a way to catch earlier and spit out more helpful messaging.

Maybe within `p_array()` check if the current.type is an identifier,
yell at them if so?