Add tests for exceptions, %include/constrain statements
Pretty huge commit. Adding a ton of tests for more exceptions, as well
as a pretty good amount for %include statements.

May want to separate %include and %constrain to different files. They
have some crossover in their concerns, but there's a pretty good amount
of complexity surrounding %constrain's. Particularly with the symbol
table merge, and what is allowed/denied.