Move semantic analysis to its own file
Begin to think about %use statements. I am of two minds.

1) Does it fit the definition of a parser statement?

Initially I had defined parser statements as something that does not
live past the parser. It is not involved in semantic analysis, and does
not end up in the compiled output.

Given this consideration, %use statements shouldn't be a "parser
statement". They are entered into the symbol table, and thus are used in
semantic analysis. They also must end up in part of the compiled output,
or the functions may not be run in the VM.

2) Do I need to rethink the definition?

Perhaps these are just statements have have an earlier "hook" into the
parser, with some added responsibilities. To be fair, FFI functions will
not end up as a DATA_$n node.

I think to fully understand the problem more, I need to think through
the design of the VM.