Get mad stuck--accomplished, refs are hard
Any solution I'm coming up with is requiring doing a ton of passes. If
we have an index of a reference:  `array[0]`, we need to resolve `array`
before we can subscript it. But if this comes before the actual
definition of the `array` variable, gotta defer compiling the result
until we've resolved `array`.

Aah.