Add notes on how to build dependency graph
It's very possible this is not worth the effort. May unnecessarily slow
down compliation, adding a number of passes to build the graph, then
another set of passes to resolve the references.

It's wasteful, and kinda ruins the legibility of the output, but it's
almost always faster to just recompile the nodes that references point
to.

If we're assuming this will mostly be used as a compiled output, it's
preferred for the output to be cleaner.

If we assume people will re-run `confc` frequently, then speed is most
important.