More progress in semantics, can't figure out indexes
Feels difficult to determine if I have an array of arrays of strings,
how can I access the type of the inner-inner strings? Need to
semi-evaluate the dumb thing to actually pull the type. E.g.,

```
things (array): [
   ["one", "two"]
];

one: things[0][0];
```

Breaks down to: `((things, 0), 0)`.

Typechecking `things[0]` is easy. But it feels like we need to access
the item within things, at index 0, to access ITS 0th element as well.