Fix +2 nested section references
A member expression such as `var: S1.S2.key` would fail, as `S1.S2`
evaluates to a section. That section was being added to the dependencies
of `var`.

Currently resolved by excluding sections from being listed as
dependencies, but it's a little inelegent. May have a better solution
down the line.